split app setup from server.js so can unit test with jest without system hanging refactored some component names because they were all App some comments added to files to help add dummyData etc
Terms And Conditions Fixed

Fixed Terms, fixed Footer

Fixed issue with the image reference
