+++ .eslintrc
@@ -6,7 +6,10 @@
   "env": {
     "browser": true,
     "node": true,
-    "mocha": true
+    "mocha": true,
+    "jasmine": true,
+    "jest": true,
+    "es6": true
   },
   "ecmaFeatures": {
     "jsx": true,
