+++ __mocks__/styleMock.js
@@ -1,0 +1,2 @@
+// workaround to get jest to play nice with express and react
+module.exports = {};
