+++ client/modules/Faq/Faq.js
@@ -2,7 +2,7 @@
 import React, { Component, PropTypes } from 'react'; 
 
 //Create a component class
-export class App extends Component {
+export class Faq extends Component {
 
   render() {
     // Here goes our page 
@@ -14,4 +14,4 @@ export class App extends Component {
   }
 }
 
-export default App;
+export default Faq;
