+++ package.json
@@ -18,6 +18,12 @@
     "lint": "eslint client server",
     "heroku-postinstall": "npm run build"
   },
+  "jest": {
+    "moduleNameMapper": {
+      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+      "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js"
+    }
+  },
   "pre-commit": [
     "lint",
     "test"
@@ -54,10 +60,12 @@
     "sanitize-html": "^1.11.4"
   },
   "devDependencies": {
+    "babel-cli": "^6.26.0",
     "babel-eslint": "^6.0.4",
     "babel-loader": "^6.2.4",
     "babel-plugin-webpack-loaders": "^0.7.0",
     "babel-polyfill": "^6.9.1",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-es2015-native-modules": "^6.6.0",
     "babel-preset-react": "^6.5.0",
@@ -79,11 +87,11 @@
     "eventsource-polyfill": "^0.9.6",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.8.5",
-    "jest": "22.4.2",
+    "jest": "^22.4.3",
     "jsdom": "^9.2.1",
     "json-loader": "^0.5.4",
     "mock-css-modules": "^1.0.0",
-    "mockgoose": "^6.0.3",
+    "mockgoose": "^6.0.8",
     "nock": "^8.0.0",
     "nodemon": "^1.9.2",
     "null-loader": "^0.1.1",
@@ -95,12 +103,14 @@
     "pre-commit": "^1.1.3",
     "react-addons-test-utils": "^15.1.0",
     "react-hot-loader": "^3.0.0-beta.2",
+    "really-need": "^1.9.2",
     "redux-devtools": "^3.3.1",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.0.11",
     "rimraf": "^2.5.2",
     "sinon": "^1.17.4",
     "style-loader": "^0.13.1",
+    "superagent": "^3.8.2",
     "supertest": "^1.2.0",
     "url-loader": "^0.5.7",
     "webpack": "2.1.0-beta.8",
