+++ server/dummyData.js
@@ -1,3 +1,11 @@
 export default function () {
+  // add dummy data here
+  // see http://mongoosejs.com/docs/queries.html for an intro on how to query mongoDb using mongoose
+  // see http://mongoosejs.com/docs/models.html for intro on how to add/save data
 
+  // add an admin user if there is not one already in the system... preferably using carshareadm@gmail.com
+
+  // add locations if there are none already in the system
+
+  // add cars if there are none already in the system
 }
