+++ server/util/test-helpers.js
@@ -1,17 +1,23 @@
 import mongoose from 'mongoose';
 import mockgoose from 'mockgoose';
+import request from 'supertest';
+import * as reallyNeed from 'really-need';
 
-export function connectDB(t, done) {
-  mockgoose(mongoose).then(() => {
-    mongoose.createConnection('mongodb://localhost:27017/mern-test', err => {
-      if (err) t.fail('Unable to connect to test database');
-      done();
+module.exports = {
+  connectDB: function (t, done) {
+    mockgoose(mongoose).then(() => {
+      mongoose.createConnection('mongodb://localhost:27017/carshare-test', err => {
+        if (err) t.fail('Unable to connect to test database');
+        done();
+      });
+    })
+  },
+  dropDB: function(t) {
+    mockgoose.reset(err => {
+      if (err) t.fail('Unable to reset test database');
     });
-  });
-}
+  },
+  request: request,
+  require: reallyNeed,
+};
 
-export function dropDB(t) {
-  mockgoose.reset(err => {
-    if (err) t.fail('Unable to reset test database');
-  });
-}
